requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
matplotlib
gunicorn
apache-airflow
requests
python-multipart
pydantic

